// Timeseries2/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
import { DataRecordValue } from '@superset-ui/core';
import { EchartsProps } from '../types';

export type TimestampType = string | number | Date;

export type EchartsBaseTimeseriesSeries = {
  name: string;
  data: [Date, DataRecordValue][];
};

export type EchartsTimeseriesSeries = EchartsBaseTimeseriesSeries & {
  color: string;
  stack?: string;
  type: 'bar' | 'line';
  smooth: boolean;
  step?: 'start' | 'middle' | 'end';
  areaStyle: {
    opacity: number;
  };
  symbolSize: number;
};

export type EchartsTimeseriesProps = EchartsProps & {
  area: number;
  colorScheme: string;
  contributionMode?: string;
  zoomable?: boolean;
  seriesType: string;
  logAxis: boolean;
  stack: boolean;
  markerEnabled: boolean;
  markerSize: number;
  minorSplitLine: boolean;
  splitLine: boolean;
  opacity: number;
  labelEnable: boolean;
  area2: number;
  colorScheme2: string;
  contributionMode2?: string;
  seriesType2: string;
  logAxis2: boolean;
  stack2: boolean;
  markerEnabled2: boolean;
  labelEnable2: boolean;
  markerSize2: number;
  minorSplitLine2: boolean;
  opacity2: number;
  labelFontSize: number;
  labelPosition: string;
};